pyproject.toml
[project]
name = "cryptoprices-bot"
version = "0.1.0"
description = "Telegram bot to get cryptocurrency prices and receive real-time price alerts"
authors = [
    {name = "Matias Rivero Hadad",email = "you@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[job-queue] (>=22.0,<23.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "aiohttp (>=3.11.16,<4.0.0)"
]


[tool.poetry]  # Required to enable groups in the new poetry-core


[tool.poetry.group.dev.dependencies]
ruff = ">=0.2.0,<0.3.0"
mypy = ">=1.7,<2.0"
pre-commit = ">=3.5,<4.0"


[tool.ruff]
target-version = "py311"
line-length = 100
fix = true
lint.select = ["E", "F", "I", "UP", "N", "B", "C4"]
lint.ignore = ["E501"]  # optional if you want to ignore line length errors


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"


[tool.mypy]
python_version = "3.11"
packages = ["src"]
check_untyped_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true
strict_optional = true
warn_unused_ignores = true
namespace_packages = true
explicit_package_bases = true


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
